Cleanup CalcArmAngleHelper
Cleanup


Cleanup


Cleanup


Cleanup


Move degrees conversions to UnitConversions.java


Add TODO
