+++ src/main/java/frc/robot/Constants.java
@@ -101,8 +101,8 @@ public final class Constants {
     public static double karmLengthFromEdgeToPivot = 0.25;
     public static double karmHeightFromWinchToPivotPoint = 0.75;
 
-    public static double karmHeightFromWinchToPivotPoint_Min = 0.1;
-    public static double karmLengthFromEdgeToPivot_Min = 0.1;
+    public static double klengthFromWinchToPivotPoint_Min = 0.1;
+    public static double klengthFromPivotPointToArmBackEnd_Min = 0.1;
     public static double karmEncoderRotationsOffset = 0.56;
 
     public static double kdeltaRotationsBeforeBroken = .01;
